+++ samples/Chat/webpack.config.js
@@ -4,6 +4,6 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 
 module.exports = (env) => {
-  const commonConfig = require('../../common/config/webpack/sample.webpack.config')(__dirname, env);
+  const commonConfig = require('../../common/config/webpack/sampleapp.webpack.config')(__dirname, env);
   return commonConfig;
 };
